requirements.txt
pandas>=1.4.0,<2.3
openpyxl>=3.1.0
PyYAML>=6.0
great_expectations>=0.17.0,<1.0
reportlab>=3.6.0
termcolor>=3.2.0